backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
httpx==0.28.0
streamlit==1.40.0
pandas==2.2.3
plotly==5.24.1
psycopg2-binary==2.9.10
